separated the server actions in another file
